+++ packages/wiz-ui-next/src/components/base/inputs/date-picker/date-picker.vue
@@ -54,7 +54,7 @@
                 as="span"
                 fontSize="xs"
                 line-height="lg"
-                color="gray.700"
+                color="gray.800"
               >
                 {{ formatYear(currentMonth.getFullYear()) }}
               </WizText>
@@ -67,7 +67,7 @@
                 >
                   <WizIArrowDropUp
                     :class="[
-                      fillStyle['gray.700'],
+                      fillStyle['gray.800'],
                       fontSizeStyle['xs2'],
                       datePickerArrowIconStyle,
                     ]"
@@ -81,7 +81,7 @@
                 >
                   <WizIArrowDropDown
                     :class="[
-                      fillStyle['gray.700'],
+                      fillStyle['gray.800'],
                       fontSizeStyle['xs2'],
                       datePickerArrowIconStyle,
                     ]"
@@ -89,7 +89,7 @@
                 </button>
               </WizVStack>
             </WizHStack>
-            <WizText as="span" fontSize="xs" color="gray.700">
+            <WizText as="span" fontSize="xs" color="gray.800">
               {{ currentDateTitle }}
             </WizText>
           </WizHStack>