+++ packages/wiz-ui-next/src/components/base/inputs/selectbox/selectbox.vue
@@ -41,7 +41,10 @@
       @onClose="openSelectBox = false"
       :isDirectionFixed="isDirectionFixed"
     >
-      <div :class="selectBoxSelectorStyle" :style="{ minWidth: width }">
+      <div
+        :class="selectBoxSelectorStyle"
+        :style="{ minWidth: width, maxHeight: dropdownMaxHeight }"
+      >
         <WizVStack gap="xs2">
           <div
             :class="selectBoxSelectorOptionStyle"
@@ -133,6 +136,10 @@ const props = defineProps({
     type: Boolean,
     required: false,
     default: false,
+  },
+  dropdownMaxHeight: {
+    type: String,
+    required: false,
   },
 });