+++ packages/wiz-ui-react/src/components/base/popup-button-group/components/button-item.tsx
@@ -99,6 +99,7 @@ export const ButtonItem: FC<Props> = ({
             <WizIcon icon={item.option.icon} color={iconColor} size="md" />
           )}
         </span>
+        {item.option.exLabel && <span>{item.option.exLabel}</span>}
       </div>
     </div>
   );