+++ packages/wiz-ui/src/components/base/buttons/icon/index.ts
@@ -1,1 +1,0 @@
-export { default as WizIconButton } from "./icon.vue";