+++ packages/wiz-ui/src/components/base/buttons/index.ts
@@ -1,4 +1,0 @@
-export * from "./text";
-export * from "./icon";
-export * from "./toggle";
-export * from "./tiny";