+++ packages/wiz-ui/src/components/base/graphs/circle/index.ts
@@ -1,1 +1,0 @@
-export { default as WizCircleGraph } from "./circle.vue";