+++ packages/wiz-ui/src/components/base/graphs/index.ts
@@ -1,6 +1,0 @@
-export * from "./circle";
-export * from "./compare";
-export * from "./poll";
-export * from "./transition";
-export * from "./meter";
-export * from "./line";