+++ packages/wiz-ui/src/components/base/icon/index.ts
@@ -1,1 +1,0 @@
-export { default as WizIcon } from "./icon.vue";