+++ packages/wiz-ui/src/components/base/information-panel/index.ts
@@ -1,1 +1,0 @@
-export { default as WizInformationPanel } from "./information-panel.vue";