+++ packages/wiz-ui/src/components/base/inputs/search-input/index.ts
@@ -1,2 +1,0 @@
-export { default as WizSearchInput } from "./search-input.vue";
-export { default as WizSearchPopup } from "./search-popup.vue";