+++ packages/wiz-ui/src/components/base/inputs/upload/index.ts
@@ -1,2 +1,0 @@
-export { default as WizUpload } from "./upload.vue";
-export { default as WizUploadDisplay } from "./upload-display.vue";