+++ packages/wiz-ui/src/components/base/tab/types.ts
@@ -1,6 +1,0 @@
-export interface TabItem {
-  label: string;
-  name: string;
-  notificationCount?: number;
-  disabled?: boolean;
-}