+++ packages/wiz-ui/src/components/custom/index.ts
@@ -1,5 +1,0 @@
-export * from "./chat";
-export * from "./form";
-export * from "./notification";
-export * from "./schedule";
-export * from "./timeline";