+++ packages/wiz-ui/src/components/custom/notification/list/index.ts
@@ -1,1 +1,0 @@
-export { default as WizNotificationList } from "./list.vue";