+++ packages/wiz-ui/src/components/manager/index.ts
@@ -1,1 +1,0 @@
-export { default as WizProvider } from "./provider.vue";