+++ packages/wiz-ui/src/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from "./use-z-index";
-export * from "./use-breakpoint";
-export { useSnackbar } from "./use-snackbar";